Refactor test code structure and clean up a bit
